small compilation fixes

+++ sql/backends/monet5/vaults/parquet/parquet.c
@@ -681,7 +681,7 @@ PARQUETread_large(BAT **R, pqc_creader *r, int colno, Pipeline *p, int wnr)
 		size_t offset = 0;
 		if (dict >= 4 && is_string)
 			offset = h->free;
-		uint64_t rsz = 0, tsz = 0;
+		int64_t rsz = 0, tsz = 0;
 		if ((rsz = pqc_read_chunk(r->c[pse->ccnr], wnr, rb->theap->base, ((char*)h->base)+h->free, sz, &offset, &dict)) < 0) {
 			BBPreclaim(rb);
 			const char *err = pqc_get_error(r->c[pse->ccnr]);
@@ -691,7 +691,7 @@ PARQUETread_large(BAT **R, pqc_creader *r, int colno, Pipeline *p, int wnr)
 		}
 		h->free += ssize;
 		tsz += rsz;
-		while(tsz < sz) {
+		while(tsz < (int64_t)sz) {
 			dict = 0;
 			if (rsz == 0)
 				break;
@@ -742,7 +742,7 @@ PARQUETread_large(BAT **R, pqc_creader *r, int colno, Pipeline *p, int wnr)
 		if (!rb) {
 			throw(SQL, "parquet.read",  SQLSTATE(HY013) MAL_MALLOC_FAIL);
 		}
-		uint64_t rsz = 0, tsz = 0;
+		int64_t rsz = 0, tsz = 0;
 		if ((rsz = pqc_read_chunk(r->c[pse->ccnr], wnr, rb->theap->base, NULL, sz, NULL, NULL)) < 0) {
 			BBPreclaim(rb);
 			const char *err = pqc_get_error(r->c[pse->ccnr]);
@@ -751,7 +751,7 @@ PARQUETread_large(BAT **R, pqc_creader *r, int colno, Pipeline *p, int wnr)
 			throw (SQL, "parquet.read", SQLSTATE(HY002) "Error reading parquet file");
 		}
 		tsz += rsz;
-		while(tsz < sz) {
+		while(tsz < (int64_t)sz) {
 			if (rsz == 0)
 				break;
 			if ((rsz = pqc_read_chunk(r->c[pse->ccnr], wnr, ((char*)rb->theap->base)+(tsz*rb->twidth), NULL, sz-tsz, NULL, NULL)) < 0) {
